int64_t to unsigned int

+++ src/chars.h
@@ -20,7 +20,7 @@ struct buf_s{
  * @param unicode
  * @return int -1:error >0: utf8单元长度
  */
-int utf8unit2unicode(const char* ptr, int64_t* unicode);
+int utf8unit2unicode(const char* ptr, unsigned int* unicode);
 
 /**
  * unicode单元转为utf8单元
@@ -29,7 +29,7 @@ int utf8unit2unicode(const char* ptr, int64_t* unicode);
  * @param bufsize
  * @return int uft8单元长度
  */
-int unicode2utf8unit(int64_t unicode, char* buf, unsigned int bufsize);
+int unicode2utf8unit(unsigned int unicode, char* buf, unsigned int bufsize);
 
 /**
  * 对字符串中的每个unicode单位应用用户回调函数
@@ -43,7 +43,7 @@ int unicode2utf8unit(int64_t unicode, char* buf, unsigned int bufsize);
 int eachUnicode(
         const char* ptr,
         unsigned int len,
-        int (*onEachUnicode)(int64_t unicode, int utf8unitlen, void* cbdata),
+        int (*onEachUnicode)(unsigned int unicode, int utf8unitlen, void* cbdata),
         void* cbdata);
 
 /**
